requirements.txt
astroid==2.0.4
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
datadog==0.26.0
dateutils==0.6.6
decorator==4.4.0
Deprecated==1.2.5
Django==2.1.5
docopt==0.6.2
docutils==0.14
dodgy==0.1.9
frosted==1.4.1
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
mypy==0.670
mypy-extensions==0.4.1
nose==1.3.7
nsetools==1.0.8
numpy==1.16.0
opencv-python==4.0.0.21
pandas==0.24.0
pbr==5.1.3
pep8-naming==0.4.1
pies==2.6.7
Pillow==5.4.1
prospector==1.1.6.2
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==1.6.0
PyGithub==1.43.7
PyJWT==1.7.1
pylint==2.1.1
pylint-celery==0.3
pylint-django==2.0.2
pylint-flask==0.5
pylint-plugin-utils==0.5
pyperclip==1.7.0
pyroma==2.4
python-dateutil==2.7.5
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requirements-detector==0.6
selenium==3.141.0
sentry-sdk==0.6.5
setoptconf==0.2.0
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.1
typed-ast==1.3.1
urllib3==1.24.1
vulture==0.24
wikipedia==1.4.0
wrapt==1.11.1
xlrd==1.2.0
